move prover/test_utils to adapter/test_utils and add adapter test mode
