+++ stwo_cairo_prover/crates/adapter/src/lib.rs
@@ -5,12 +5,13 @@ use memory::Memory;
 use opcodes::StateTransitions;
 use serde::{Deserialize, Serialize};
 
+pub mod adapter;
 pub mod builtins;
 pub mod decode;
 pub mod memory;
 pub mod opcodes;
-pub mod plain;
 pub mod relocator;
+pub mod test_utils;
 pub mod vm_import;
 
 pub const N_REGISTERS: usize = 3;
