+++ stwo_cairo_prover/crates/prover/src/lib.rs
@@ -7,6 +7,3 @@ pub mod prover;
 
 pub use stwo_prover;
 pub mod witness;
-
-#[cfg(test)]
-mod test_utils;
